Moved constants to constant file
